+++ LightwaveRF.h
@@ -12,11 +12,11 @@
 #define LW_MOOD 2
 #define LW_LOCK 4
 
-extern void lw_tx_setup(int tx_pin);
+extern void lw_tx_setup(byte tx_pin);
 
-extern void lw_rx_setup(int rx_pin, int interrupt);
+extern void lw_rx_setup(byte rx_pin, byte interrupt);
 
-extern boolean lw_setup(int tx_pin, int rx_pin, int interrupt);
+extern boolean lw_setup(byte tx_pin, byte rx_pin, byte interrupt);
 
 extern void lw_setup();
